# openapi.yaml
openapi: 3.0.3
info:
  title: Non-Linear Video Editor API
  description: |
    API for managing video editing projects, generating AI content (videos, images, audio),
    and handling user subscriptions.

    ## Authentication
    All endpoints require session-based authentication via Supabase Auth cookies.

    ## Rate Limiting
    The API uses a tiered rate limiting system:
    - Tier 1: 5 req/min (auth, payments)
    - Tier 2: 10 req/min (resource creation)
    - Tier 3: 30 req/min (read operations)
    - Tier 4: 60 req/min (general)
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: https://your-domain.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: supabase-auth-token
      description: Supabase authentication session cookie
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        field:
          type: string
          description: Field that caused the error
    RateLimitError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            limit:
              type: integer
              description: Rate limit maximum
            remaining:
              type: integer
              description: Remaining requests
            resetAt:
              type: integer
              format: int64
              description: Timestamp when limit resets
...
    Asset:
      type: object
      required:
        - id
        - project_id
        - user_id
        - type
        - storage_url
        - created_at
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [image, video, audio]
        source:
          type: string
          enum: [upload, genai]
        storage_url:
          type: string
          description: Internal storage URL
        mime_type:
          type: string
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
...
    VideoStatusResponse:
      oneOf:
        - type: object
          required:
            - done
            - progress
          properties:
            done:
              type: boolean
              enum: [false]
            progress:
              type: number
              minimum: 0
              maximum: 100
        - type: object
          required:
            - done
            - asset
            - storageUrl
          properties:
            done:
              type: boolean
              enum: [true]
            asset:
              $ref: '#/components/schemas/Asset'
            storageUrl:
              type: string
              format: uri
        - type: object
          required:
            - done
            - error
          properties:
            done:
              type: boolean
              enum: [true]
            error:
              type: string
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - cookieAuth: []
paths:
...
  /video/status:
    get:
      tags:
        - Video
      summary: Check video generation status
      description: Poll the status of a video generation operation
      operationId: getVideoStatus
      parameters:
        - name: operationName
          in: query
          required: true
          schema:
            type: string
          description: Operation ID from /video/generate
        - name: projectId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoStatusResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
...